Update Exercice 2.ino

+++ Exercice 2.ino
@@ -3,10 +3,10 @@ void setup(){
 }
 
 void loop(){
-    int nombre;
+    float nombre;
     if (Serial.available()){
         nombre = Serial.parseInt ();
         nombre = sq(nombre);
         Serial.println(nombre);
     }
-} 
+}
